# perf: speed up access to "uo_out" while generating image from the sim