Caching the index of "uo_out" speeds up the "stripes" by almost 40% from 50fps (20ms) to 77fps (13ms). Heavy "Drop" example is 30% faster from 32fps (31ms) to 42fps (24ms).